+++ builder_pkgs/mockito/test/builder/auto_mocks_test.dart
@@ -808,6 +808,40 @@ void main() {
     );
   });
 
+  test('generates mock classes from an annotation on an import directive',
+      () async {
+    var mocksOutput = await buildWithNonNullable({
+      ...annotationsAsset,
+      'foo|lib/foo.dart': dedent(r'class Foo {} class Bar {}'),
+      'foo|test/foo_test.dart': '''
+        @GenerateMocks([Foo])
+        import 'package:foo/foo.dart';
+        import 'package:mockito/annotations.dart';
+        '''
+    });
+    expect(mocksOutput,
+        contains('class MockFoo extends _i1.Mock implements _i2.Foo'));
+  });
+
+  test('generates mock classes from an annotation on an export directive',
+      () async {
+    var mocksOutput = await buildWithNonNullable({
+      ...annotationsAsset,
+      'foo|lib/foo.dart': dedent(r'''
+        class Foo {}
+        class Bar {}
+        '''),
+      'foo|test/foo_test.dart': '''
+        @GenerateMocks([Foo])
+        export 'dart:core';
+        import 'package:foo/foo.dart';
+        import 'package:mockito/annotations.dart';
+        '''
+    });
+    expect(mocksOutput,
+        contains('class MockFoo extends _i1.Mock implements _i2.Foo'));
+  });
+
   test('generates multiple mock classes', () async {
     var mocksOutput = await buildWithNonNullable({
       ...annotationsAsset,